requirements.txt
datasets
tensorflow
numpy
pillow
matplotlib
scikit-learn
opencv-python
gradio
huggingface-hub
tqdm
requests